moved foodlist functionality to foodlist.js
