+++ frontend/src/components/Panels/TopSubPanel.js
@@ -8,7 +8,6 @@ import {getDateString} from "../divHelpers/monthGenerator";
 const storage = require("../tokenStorage.js");
 
 function TopSubPanel(props){
-  console.log(props.date)
     function makeRemoveFoodJSON(id){
       const removeData = {
         foodRecordId: id,
@@ -16,37 +15,41 @@ function TopSubPanel(props){
       }
       return JSON.stringify(removeData);
     }
-    
-    async function removeFood(id, index){    
-      const deleteJSON = makeRemoveFoodJSON(id);
+ function makeFoodButtons(id, index){
 
-      console.log(deleteJSON);
-    
-      let res = await JSONRequest("DELETE", deleteJSON, "api/users/data/foodRecords");
-      
-      console.log(res.error);
-    
-      foods.splice(index,1);
-      setFoods(foods.slice(0));
-    }
-    
-    function makeFoodButtons(id, index){
-    
         return(
             <div className= "buttons">
               {makeActionButton("button","removeFoodButton",() => removeFood(id, index),"x",id)}
-              {makeActionButton("button","removeFoodButton",() => editFood(id, index),"&#9998;",id)}
             </div>
         )
     }
-    
-    function FoodElement(props){
+
+
+
+
+    async function removeFood(id, index){
+      const deleteJSON = makeRemoveFoodJSON(id);
+
+      console.log(deleteJSON);
+
+      let res = await JSONRequest("DELETE", deleteJSON, "api/users/data/foodRecords");
+      props.updateFoods();
+
+    }
+
+
+    function FoodList(props){
+      return(
+        props.foods.map((f, index) => <FoodElement key={f.id} food={f} index={index}/>)
+      )
+    }
+  function FoodElement(props){
       let foodEl = props.food.food;
       let portion = props.food.amountConsumed.portion;
       let quantity = props.food.amountConsumed.quantity;
       console.log(portion);
       return(
-    
+
         <div className = "dayFood" key={foodEl.id} >
           <div className ="foodName">
             {foodEl.description}
@@ -64,66 +67,10 @@ function TopSubPanel(props){
         </div>
       )
     }
-    
-    
-    function FoodList(props){
-      return(
-        props.foods.map((f, index) => <FoodElement key={f.id} food={f} index={index}/>)
-      )
-    }
-
-    function makeFoodDayJSON(curDate){
-      const dateString = getDateString(curDate);
-      const foodReq = {
-        userId:props.userId,
-        startDate:dateString,
-        endDate:dateString,
-        jwtToken:storage.retrieveToken()
-      }
-      return foodReq;
-    }
-
-    function makeFoodDayURL(foodReq){
-      return "api/users/data/foodRecords/?userId="+foodReq.userId+"&startDate="+foodReq.startDate+"&endDate="+foodReq.endDate+"&jwtToken="+foodReq.jwtToken;
-    }
-
-    function handleFDGetRes(res){
-      if(res.error!=0){
-        console.log("error Happened"+res.error);
-        return [];
-      }
-      console.log(res.foodRecords);
-      console.log("no error");
-      return res.foodRecords
-    }
-
-    
-    // this will get the latest version of the person's food day
-    async function getFoodDayList(startDate){
-      console.log(makeFoodDayURL(makeFoodDayJSON(startDate)));
-      let res = await JSONGETRequest(makeFoodDayURL(makeFoodDayJSON(startDate)));
-      console.log(res);
-      return handleFDGetRes(res);
-    }
-
-    const [fl,setFl] = useState(props.foodList);
-    const [foods,setFoods] = useState([]);
-    // gets initial food day data
-    useEffect(() =>{
-      const getRecords = async () =>{
-        let res = await getFoodDayList(props.date);
-        setFoods(res);
-        console.log("hello");
-        console.log(res);
-      }
-      getRecords();
-    },[]);
-
-    console.log(foods);
 
     return(
       <div id = "topSubPanel">
-        <FoodList foods ={foods}/>
+        <FoodList foods ={props.foods}/>
       </div>
     )
 }
